fix(cuda): accumulate half cuDNN convolutions in float

+++ candle-core/src/cuda_backend/mod.rs
@@ -1900,7 +1900,10 @@ impl CudaStorage {
                 let inp = &inp.slice(inp_l.start_offset()..);
                 let k = &k.slice(kernel_l.start_offset()..);
                 let mut out = unsafe { device.alloc::<f16>(dst_el)? };
-                crate::cudnn::launch_conv1d::<f16, f16>(inp, inp_l, k, &mut out, params, &device)
+                // Half storage, float accumulation: PyTorch v2.5.1
+                // aten/src/ATen/cudnn/Descriptors.h:202-205. Tensor-op math
+                // alone does not promote the convolution compute descriptor.
+                crate::cudnn::launch_conv1d::<f16, f32>(inp, inp_l, k, &mut out, params, &device)
                     .map_err(crate::Error::wrap)?;
                 S::F16(out)
             }
@@ -2028,7 +2031,10 @@ impl CudaStorage {
                 let inp = &inp.slice(inp_l.start_offset()..);
                 let k = &k.slice(kernel_l.start_offset()..);
                 let mut out = unsafe { device.alloc::<f16>(dst_el)? };
-                crate::cudnn::launch_conv2d::<f16, f16>(inp, inp_l, k, &mut out, params, &device)
+                // Half storage, float accumulation: PyTorch v2.5.1
+                // aten/src/ATen/cudnn/Descriptors.h:202-205. Tensor-op math
+                // alone does not promote the convolution compute descriptor.
+                crate::cudnn::launch_conv2d::<f16, f32>(inp, inp_l, k, &mut out, params, &device)
                     .map_err(crate::Error::wrap)?;
                 S::F16(out)
             }
